+++ src/apps/shared/hooks.ts
@@ -40,3 +40,34 @@ export function useTimeout(cb: () => void, ms: number, deps: any[] = []) {
     return clearLastTimeout;
   }, [ms, ...deps]);
 }
+
+export function useSyncClockInterval(cb: () => void, on: 'minutes' | 'seconds', deps: any[] = []) {
+  const ref = useRef<number | null>(null);
+
+  const clearLastInterval = () => {
+    if (ref.current) {
+      clearInterval(ref.current);
+    }
+  };
+
+  useEffect(() => {
+    clearLastInterval();
+
+    const now = new Date();
+    let msToWaitForClockSync = 0;
+    if (on === 'minutes') {
+      const secondsUntilNextMinute = 60 - now.getSeconds();
+      msToWaitForClockSync = secondsUntilNextMinute * 1000 - now.getMilliseconds();
+    } else if (on === 'seconds') {
+      msToWaitForClockSync = 1000 - now.getMilliseconds();
+    }
+
+    setTimeout(() => {
+      cb();
+      let interval = on === 'minutes' ? 60 * 1000 : 1000;
+      ref.current = window.setInterval(cb, interval);
+    }, msToWaitForClockSync);
+
+    return clearLastInterval;
+  }, [on, ...deps]);
+}